fix edit avatar cust omer

+++ src/models/customer.ts
@@ -156,12 +156,11 @@ export default class CustomerModel {
     placeOfDelivery: string,
     bio: string,
     user: User,
-    file: Express.Multer.File
+    file: Express.Multer.File | null
   ) {
     // find customer on database
     const customerRepository = ShopPDataSource.getRepository(Customer);
     const localFileRepository = ShopPDataSource.getRepository(LocalFile);
-
     if (user.customer == null)
       return new Response(
         HttpStatusCode.REDIRECT,
@@ -179,21 +178,20 @@ export default class CustomerModel {
         bio,
       }
     );
-
-    const localFileEdit = await localFileRepository.update(
-      {
-        id: user.customer.avatar.id,
-      },
-      {
-        filename: file.filename,
-        mimetype: file.mimetype,
-        path: file.path,
-      }
-    );
-
-    deleteFile(user.customer.avatar.path);
-
-    if (result.affected == 1 && localFileEdit.affected == 1) {
+    if (file != null) {
+      const localFileEdit = await localFileRepository.update(
+        {
+          id: user.customer.avatar.id,
+        },
+        {
+          filename: file.filename,
+          mimetype: file.mimetype,
+          path: file.path,
+        }
+      );
+      deleteFile(user.customer.avatar.path);
+    }
+    if (result.affected == 1) {
       return new Response(HttpStatusCode.OK, 'Edit customer successfully!');
     } else {
       return new Response(HttpStatusCode.BAD_REQUEST, 'Edit customer failed !');
